# server/Webby.RoomService/docs/swagger.yaml
basePath: /
definitions:
  docs.ApiResponse-PaginatedResponse[RoomListItem]:
    properties:
      data:
        $ref: '#/definitions/docs.PaginatedResponse-docs_RoomListItem'
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  docs.ApiResponse-docs_RoomResponse:
    properties:
      data:
        $ref: '#/definitions/docs.RoomResponse'
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  docs.Error400Response:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          name: Name must be at least 2 characters
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  docs.Error401Response:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          message: Missing or invalid token
        type: object
      message:
        example: Unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
  docs.Error403Response:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          message: Admin only
        type: object
      message:
        example: Forbidden
        type: string
      success:
        example: false
        type: boolean
    type: object
  docs.Error404Response:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          message: Object not found
        type: object
      message:
        example: Not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  docs.Error500Response:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          message: Database connection lost
        type: object
      message:
        example: Internal server error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  docs.PaginatedResponse-docs_RoomListItem:
    properties:
      items:
        items:
          $ref: '#/definitions/docs.RoomListItem'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  docs.RoomListItem:
    properties:
      categoryId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      hostId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isPrivate:
        example: false
        type: boolean
      name:
        example: Room Name
        type: string
      thumbnail:
        example: https://example.com/thumbnail.jpg
        type: string
    type: object
  docs.RoomResponse:
    properties:
      categoryId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      hostId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isPrivate:
        example: false
        type: boolean
      name:
        example: Room Name
        type: string
      thumbnail:
        example: https://example.com/thumbnail.jpg
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This API provides endpoints for managing rooms and categories.
  title: Webby.RoomService
  version: "1.0"
paths:
...
  /rooms/{id}:
    delete:
      description: |-
        Delete a room by ID. **Permanent action - cannot be undone.**
        **Path Parameter Validation:**
        * `id`: must be a valid UUID v4 format
        **Security Note:**
        * This action **requires authentication**
        * **Only the authorized room creator** can delete this room
      parameters:
      - description: Room ID (UUID v4 format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Room successfully deleted - No Content returned
        "400":
          description: Invalid UUID format in path
          schema:
            $ref: '#/definitions/docs.Error400Response'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/docs.Error401Response'
        "403":
          description: Not authorized - only creator or admin can delete
          schema:
            $ref: '#/definitions/docs.Error403Response'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/docs.Error404Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.Error500Response'
      security:
      - BearerAuth: []
      summary: Delete a room
      tags:
      - Rooms
    get:
      description: |-
        Retrieve a room by its ID.
        **Access Rules:**
        * **Public rooms** can be retrieved by any authenticated user
        * **Private rooms** can only be retrieved by the room creator or administrators
        **Path Parameter Validation:**
        * `id`: must be a valid UUID v4 format
        **Security Note:**
        * This action **requires authentication**
      parameters:
      - description: Room ID (UUID v4 format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room successfully retrieved
          schema:
            $ref: '#/definitions/docs.ApiResponse-docs_RoomResponse'
        "400":
          description: Invalid UUID format in path
          schema:
            $ref: '#/definitions/docs.Error400Response'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/docs.Error401Response'
        "403":
          description: Access denied - private room and not owner
          schema:
            $ref: '#/definitions/docs.Error403Response'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/docs.Error404Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.Error500Response'
      security:
      - BearerAuth: []
      summary: Get a room by ID
      tags:
      - Rooms
    put:
      consumes:
      - multipart/form-data
      description: |-
        Update a room's name, **category ID**, privacy setting, and **thumbnail** by ID.
        **Validation Rules:**
        * `name`: **required**, 2-50 characters, cannot be null/empty/whitespace-only
        * `categoryId`: **required**, must be a valid UUID v4 format
        * `isPrivate`: **required**, must be valid boolean ("true" or "false")
        * `thumbnail`: optional file upload, max 2MB
        **Path Parameter Validation:**
        * `id`: must be a valid UUID v4 format
        **Security Note:**
        * This action **requires authentication**
        * **Only the authorized room creator** can update this room
      parameters:
      - description: Room ID (UUID v4 format)
        in: path
        name: id
        required: true
        type: string
      - description: Room name (2-50 chars, required)
        in: formData
        maxLength: 50
        minLength: 2
        name: name
        required: true
        type: string
      - description: Category ID (UUID v4, required)
        format: uuid
        in: formData
        name: categoryId
        required: true
        type: string
      - description: Visibility flag ('true' or 'false', required)
        in: formData
        name: isPrivate
        required: true
        type: string
      - description: New thumbnail image (optional, max 2MB)
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Room successfully updated
          schema:
            $ref: '#/definitions/docs.ApiResponse-docs_RoomResponse'
        "400":
          description: 'Invalid input: name validation failed, invalid UUID, invalid
            boolean, file too large, or missing required fields'
          schema:
            $ref: '#/definitions/docs.Error400Response'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/docs.Error401Response'
        "403":
          description: Not authorized - only creator or admin can update
          schema:
            $ref: '#/definitions/docs.Error403Response'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/docs.Error404Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.Error500Response'
      security:
      - BearerAuth: []
      summary: Update a room
      tags:
      - Rooms
  /rooms/my:
    get:
      description: |-
        Retrieve a paginated list of all rooms created by the authenticated user.
        **Query Parameter Validation:**
        * `page`: **optional** (default: 1), must be >= 1, must be numeric
        * `limit`: **optional** (default: 10), must be 1-100, must be numeric
        **Security Note:**
        * This action **requires authentication**
        * Returns **only** the authenticated user's rooms
        * List includes both public and private rooms owned by the user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User's rooms successfully retrieved
          schema:
            $ref: '#/definitions/docs.ApiResponse-PaginatedResponse[RoomListItem]'
        "400":
          description: 'Invalid query parameters: page or limit not numeric, or outside
            valid range'
          schema:
            $ref: '#/definitions/docs.Error400Response'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/docs.Error401Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.Error500Response'
      security:
      - BearerAuth: []
      summary: List authenticated user's rooms
      tags:
      - Rooms
  /rooms/public:
    get:
      description: |-
        Retrieve a paginated list of all publicly visible rooms in the system.
        **Query Parameter Validation:**
        * `page`: **optional** (default: 1), must be >= 1, must be numeric
        * `limit`: **optional** (default: 10), must be 1-100, must be numeric
        **Security Note:**
        * This action **requires authentication**
        * Returns **only** public rooms (private rooms are excluded)
        * Pagination prevents excessive data retrieval
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Public rooms successfully retrieved
          schema:
            $ref: '#/definitions/docs.ApiResponse-PaginatedResponse[RoomListItem]'
        "400":
          description: 'Invalid query parameters: page or limit not numeric, or outside
            valid range'
          schema:
            $ref: '#/definitions/docs.Error400Response'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/docs.Error401Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.Error500Response'
      security:
      - BearerAuth: []
      summary: List all public rooms
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    description: Type "**Bearer**" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"